feat(results): add new statistics on activities and exercises
